+++ src/main/java/com/propadda/prop/dto/ResidentialPropertyRequest.java
@@ -12,7 +12,7 @@ public class ResidentialPropertyRequest {
 
     private String description;
 
-    private Integer price;
+    private Long price;
 
     private Integer maintenance;
 
@@ -170,11 +170,11 @@ public class ResidentialPropertyRequest {
         this.description = description;
     }
 
-    public Integer getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
